# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OMPAYE",
        "description": "API pour la gestion des comptes et transactions OMPAYE",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://ompaye-api.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/v1/accounts/{numerocompte}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Informations du compte",
                "description": "Récupérer les informations détaillées d'un compte par numéro",
                "operationId": "getAccount",
                "parameters": [
                    {
                        "name": "numerocompte",
                        "in": "path",
                        "description": "Numéro du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "KC-00001234"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations du compte récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/accounts/{numerocompte}/balance": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Solde du compte",
                "description": "Récupérer uniquement le solde d'un compte",
                "operationId": "getAccountBalance",
                "parameters": [
                    {
                        "name": "numerocompte",
                        "in": "path",
                        "description": "Numéro du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "KC-00001234"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "numerocompte": {
                                            "type": "string",
                                            "example": "KC-00001234"
                                        },
                                        "solde": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 12500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/accounts/{numerocompte}/transactions": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Transactions du compte",
                "description": "Récupérer la liste paginée des transactions d'un compte",
                "operationId": "getAccountTransactions",
                "parameters": [
                    {
                        "name": "numerocompte",
                        "in": "path",
                        "description": "Numéro du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "KC-00001234"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de la page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 45
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountResponse": {
                "title": "Réponse compte",
                "description": "Informations détaillées d'un compte",
                "properties": {
                    "numerocompte": {
                        "type": "string",
                        "example": "KC-00001234"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+22177xxxxxxx"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 12500
                    },
                    "qrCode": {
                        "description": "QR Code en base64",
                        "type": "string",
                        "example": "data:image/png;base64,..."
                    },
                    "transactions_summary": {
                        "properties": {
                            "count": {
                                "type": "integer",
                                "example": 12
                            },
                            "last": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-11-10T14:22:00Z"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "title": "Transaction",
                "description": "Modèle représentant une transaction",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "compte_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "merchant_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "depot",
                            "retrait",
                            "paiement",
                            "transfert"
                        ],
                        "example": "depot"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "en_attente",
                            "valide",
                            "annule"
                        ],
                        "example": "valide"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le token JWT dans le format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}